requirements.txt
flask
langchain
langchain-community
sentence-transformers
faiss-cpu
lxml
python-dotenv
langchain-deepseek
prompt-toolkit
Pillow
PyJWT
requests
